requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
exceptiongroup==1.2.0
Faker==22.0.0
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
kaitaistruct==0.10
magic-filter==1.0.12
multidict==6.0.4
outcome==1.3.0.post0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
russian-names==0.1.2
selenium==4.16.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
wsproto==1.2.0
yarl==1.9.4
zstandard==0.22.0
